requirements.txt
# Dev environment
pip
autopep8

# App
flask
google-generativeai
python-dotenv
markdown2

# RAG Multimodal System
google-cloud-aiplatform
vertexai
numpy
pandas
Pillow
PyMuPDF
rich
